new_corpus(2015.5.10)

+++ test/test/mltk/task/t_mcmf/T_MCMFTrSentimentModelTest.java
@@ -2,6 +2,11 @@ package test.mltk.task.t_mcmf;
 
 import org.mltk.task.t_mcmf.T_MCMFTrSentimentModel;
 
+/**
+ * 
+ * @author superhy
+ *
+ */
 public class T_MCMFTrSentimentModelTest {
 
 	public static void main(String[] args) {
@@ -10,7 +15,7 @@ public class T_MCMFTrSentimentModelTest {
 		String TEST_CORPUS_PATH = ".\\file\\sentiment\\seg\\dianziall";
 		int topicNum = 50;
 		int genWordNum = 80;
-		int lenda = 0;
+		double lenda = 0.0;
 
 		T_MCMFTrSentimentModel sentimentAnalysis = new T_MCMFTrSentimentModel(
 				TRAIN_CORPUS_PATH, TEST_CORPUS_PATH);
